Dev picking

Approved-by: Deepti Verma

+++ pos/offers.py
@@ -79,7 +79,7 @@ class BasicCartOffers(object):
                 coupon = offers_mapping[
                     product_map.retailer_product.id] if product_map.retailer_product.id in offers_mapping else {}
                 # Add/remove/update combo on a product
-                offers_list = BasicCartOffers.basic_combo_offers(int(product_map.qty), float(product_map.selling_price),
+                offers_list = BasicCartOffers.basic_combo_offers(float(product_map.qty), float(product_map.selling_price),
                                                                  product_map.retailer_product.id, coupon, offers_list)
         return offers_list
